requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
APScheduler==3.10.4
python-dotenv==1.0.0
